move All manifest sections into separate components. Including a TsdfSection, GeneratorSection, WasteLineSection, TransporterSection
In addition:
add pyupgrade pre-commit hook and upgrade hooks
