+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ ci:
 
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: 'v3.2.0'
+    rev: 'v4.5.0'
     hooks:
       - id: trailing-whitespace
       - id: check-json
@@ -24,7 +24,7 @@ repos:
             \.vscode\/.*|
           )$
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: 'v0.1.3'
+    rev: 'v0.1.11'
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
@@ -38,13 +38,17 @@ repos:
           - python
         pass_filenames: false
   - repo: https://github.com/pre-commit/mirrors-eslint
-    rev: 'v8.34.0'
+    rev: 'v8.56.0'
     hooks:
       - id: eslint
         files: \.[jt]sx?$
         types: [file]
   - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: 'v3.0.0-alpha.3'
+    rev: 'v3.1.0'
     hooks:
       - id: prettier
         exclude: infra/haztrak/
+  - repo: https://github.com/asottile/pyupgrade
+    rev: v3.15.0
+    hooks:
+      - id: pyupgrade
