+++ client/src/components/Manifest/GeneralInfo/index.ts
@@ -1,0 +1,1 @@
+export { GeneralInfoForm } from './GeneralInfoForm';
