+++ client/src/components/Manifest/Generator/GeneratorForm.spec.tsx
@@ -1,10 +1,10 @@
 import '@testing-library/jest-dom';
 import { fireEvent } from '@testing-library/react';
-import { GeneratorForm } from 'components/Manifest/Handler/GeneratorForm';
 import { siteType } from 'components/Manifest/manifestSchema';
 import React from 'react';
 import { cleanup, renderWithProviders, screen } from 'test-utils';
 import { afterEach, describe, expect, test } from 'vitest';
+import { GeneratorForm } from './GeneratorForm';
 
 afterEach(() => {
   cleanup();
