+++ client/src/components/Manifest/Generator/GeneratorForm.tsx
