+++ client/src/components/Manifest/Generator/index.ts
@@ -1,0 +1,1 @@
+export { GeneratorSection } from './GeneratorSection';
