+++ client/src/components/Manifest/Handler/index.ts
@@ -1,7 +1,6 @@
 import { Handler, handlerSchema, Signer } from 'components/Manifest/manifestSchema';
-import { GeneratorForm } from 'components/Manifest/Handler/GeneratorForm';
-import { HandlerSearchForm } from './Search/HandlerSearchForm';
 import { AddHandler } from './AddHandler';
+import { HandlerSearchForm } from './Search/HandlerSearchForm';
 
-export { GeneratorForm, HandlerSearchForm, AddHandler, handlerSchema };
+export { HandlerSearchForm, AddHandler, handlerSchema };
 export type { Handler, Signer };
