+++ client/src/components/Manifest/Tsdf/index.ts
@@ -1,0 +1,1 @@
+export { TsdfSection } from './TsdfSection';
